Finished IP packaging script, but still a few things to take care of
In the very near future I need a way to enable/disable AXI Stream
side channels with parameters. It owuld be nice if I could somehow
use the automatic widths and stuff that Xilinx IP cores offer.

Also, I need to rename the bus interfaces to not be "in" or "out".
It doesn't break anything, but it causes a very annoying critical
warning in Vivado. Well, it's a good thing Vivado warns you instead
of silently letting you deal with obscure errors late in the place
and route stage